Refactor configuration management and update launch settings
- Renamed the launch configuration from "Launch ThirdEye" to "Launch PhoenixEpsilon" for better alignment with project branding.
- Removed outdated configuration files (`config.ts` and `configLoader.ts`) to streamline the codebase and improve maintainability.
- Introduced a new configuration loading mechanism in `src/core/config/configLoader.ts`, enhancing the structure and validation of configuration settings.
- Updated the death message listener to utilize the new configuration loader, ensuring consistent access to configuration values.

These changes enhance the overall organization and robustness of the configuration management system, facilitating future enhancements and maintaining brand consistency.
